refactored health check classes
